RavenDB-15742 moved name uniqueness test to a dedicated class
