+++ test/SlowTests/Server/Documents/ETL/Olap/LocalTests.cs
@@ -33,42 +33,6 @@ namespace SlowTests.Server.Documents.ETL.Olap
         }
 
         [Fact]
-        public void NameUniqueness()
-        {
-            var script = @"
-var orderDate = new Date(this.OrderedAt);
-var year = orderDate.getFullYear();
-var month = orderDate.getMonth();
-var key = new Date(year, month);
-
-loadToOrders(partitionBy(key),
-    {
-        Company : this.Company,
-        ShipVia : this.ShipVia
-    });
-";
-
-            var path = GetTempPath("Orders");
-
-            using (var store = GetDocumentStore())
-            {
-                SetupLocalOlapEtl(store, script, path, name: null);
-                SetupLocalOlapEtl(store, script, path, name: null);
-
-                var record = store.Maintenance.Server.Send(new GetDatabaseRecordOperation(store.Database));
-                Assert.Equal(2, record.OlapEtls.Count);
-
-                var name1 = record.OlapEtls[0].Name;
-                var name2 = record.OlapEtls[1].Name;
-
-                Assert.NotEqual(name1, name2);
-                Assert.Contains(name1, name2);
-                Assert.Equal($"{name1} #2", name2);
-
-            }
-        }
-
-        [Fact]
         public Task SimpleTransformation()
         {
             var script = @"
@@ -84,7 +48,6 @@ loadToOrders(partitionBy(key),
     });
 ";
             return SimpleTransformationInternal(script);
-
         }
 
         [Fact]
